Updated for 3.0.
Added in check for treeview plugin config.

Tons of fixes, allowing for output passthrough, taking advantage of installed terminals, better control over stack, and better feedback when debugger is running.

Tons of bug fixes and little things.

Cleaned up debugger watch window.

Added in saving for build settings like arguments, as well as saving of debugging breakpoints.

Builds a project when binary not present if you try to debug.

Swapped over to new storage system from 3.0.

Added in ability to skip files; automatically skips anything in /usr/include.

Added in select-target and variants.

Added in explanation of internal, and fixed a typo.

Allowed for shells and terminals to be specified with more precision.

Added in persist.